fix(trackers): harden tracking UI and repair MangaBaka/MangaUpdates/Hikka
UI: swap +/- stepper positions, stop rating flashing 100 (blank
placeholder, wheel index-0 fallback, smiley/MU unset dash), longer
tracker-update snackbar (Short->Long), compact dialog (520->400dp).

Backends: MangaBaka refresh-fail logout, fractional ratings, null-safe
DTOs, unauth series details, no-throw status maps; MangaUpdates add
dead-code fix, 404-only rating null; Hikka OAuth deep-link wiring,
login error propagation, guarded slug parsing.

+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaScreenModel.kt
@@ -1334,7 +1334,7 @@ class MangaScreenModel(
             val result = snackbarHostState.showSnackbar(
                 message = context.stringResource(MR.strings.confirm_tracker_update, maxChapterNumber.toInt()),
                 actionLabel = context.stringResource(MR.strings.action_ok),
-                duration = SnackbarDuration.Short,
+                duration = SnackbarDuration.Long,
                 withDismissAction = true,
             )
